Fix number convert y precision y scale en los campos decimal

+++ proyecto/db/migrate/20170829040439_remove_amountwithouttax_to_accountreceivables.rb
@@ -1,5 +1,5 @@
 class RemoveAmountwithouttaxToAccountreceivables < ActiveRecord::Migration
   def change
-    remove_column :accountreceivables, :amountWithoutTax, :float
+    remove_column :accountreceivables, :amountWithoutTax, :decimal
   end
 end
