+++ proyecto/db/migrate/20170829040700_remove_totalretentions_to_accountreceivables.rb
@@ -1,5 +1,5 @@
 class RemoveTotalretentionsToAccountreceivables < ActiveRecord::Migration
   def change
-    remove_column :accountreceivables, :totalRetentions, :float
+    remove_column :accountreceivables, :totalRetentions, :decimal
   end
 end
