+++ proyecto/db/migrate/20170829040814_remove_totalamountperceive_to_accountreceivables.rb
@@ -1,5 +1,5 @@
 class RemoveTotalamountperceiveToAccountreceivables < ActiveRecord::Migration
   def change
-    remove_column :accountreceivables, :totalAmountPerceive, :float
+    remove_column :accountreceivables, :totalAmountPerceive, :decimal
   end
 end
